+++ packages/theme/src/theme/index.ts
@@ -1,5 +1,4 @@
 export * from './border'
-export * from './color.utils'
 export * from './colors'
 export * from './css-vars'
 export * from './theme'