+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @rasreee/theme
 
+## 6.0.0
+
+### Major Changes
+
+- remove @emotion from root workspace
+
+## 5.0.0
+
+### Major Changes
+
+- move and add color utils
+
+## 4.0.0
+
+### Major Changes
+
+- generated css var object from theme
+
 ## 3.0.0
 
 ### Major Changes